Integrate RPC failover into NetworkController
When configuring an Infura network, a list of failover endpoint URLs can
now be supplied. If Infura is perceived to be down (after making 15
attempts to make a request, all of which either throw an error, return a
non-200 response, or return a non-parseable response) then the request
will be forwarded automatically to the first failover URL (using
subsequent URLs as failovers for previous URLs). By default, all Infura
networks will use Quicknode as a failover.

Under the hood, the list of failover URLs are loaded into an
RpcServiceChain (added in a previous commit) and this is then passed to
the Infura middleware.