feat: update notification services to support mobile. (#5120)
## Explanation

This is a large PR that enables support for push notifications on
mobile.

**Makes Push Notification Controller Platform Agnostic**

The `NotificationServicesPushController` was highly tied to web and was
not compatible for react-native.
We now have isolated the web logic in`web/push-utils.ts`, and allow
platforms to overwrite and inject a push service into the controller.
E.g. Mobile can inject push services using react-native modules.

**Add support to toggle push notifications on/off in isolation**

This allows us to decouple in-app notifications from push notifications.
It is done by adding a `isPushEnabled` boolean inside the
`NotificationServicesPushController`.

We have also added public methods `enablePushNotifications` and
`disablePushNotifications` inside the `NotificationServicesController`
to allow us tie it to UI Actions.

## References

## Changelog

### `@metamask/notification-services-controller`

- **ADDED**: `isPushEnabled` state to
`NotificationServicesPushControllerState` to enable and disable this
controller/
- **ADDED**: `isUpdatingFCMToken` state to
`NotificationServicesPushControllerState` to track when the controller
is updating for firebase registration token.
- **ADDED**: `PushService` interface and default web implementation for
push notifciations
- **CHANGED (BREAKING)**: `NotificationServicesPushController` config
now allows injecting of a `PushService` interface during controller
creation.
- **ADDED**: `/push-services/web` subpath export to import web specific
push services.
- **ADDED**: `/shared` folder including some shared utils used in
controllers for this package.
- **ADDED**: public method `enablePushNotifications` and
`disablePushNotifications` in `NotificationServicesController` to enable
and disable push notifications in isolation.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes