test(accounts-controller): remove duplicated test mock function (#5387)
## Explanation

We had a duplicated function to create mocked account. One of them was
not handling the new `scopes` properly. So this is just a refactor to
merge both of these functions together while also fixing the logic for
future-PRs to come.

## References

N/A

## Changelog

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes