Integrate RPC failover into NetworkController (#5360)
When configuring a network, a list of failover endpoint URLs can now be
supplied. If the network is perceived to be down (after 15 unsuccessful
attempts to make a request) then the request will be forwarded
automatically to the first failover URL (using subsequent URLs as
failovers for previous URLs).

Under the hood, the list of failover URLs are loaded into an
RpcServiceChain (added in a previous commit) and this is then passed to
the Infura or fetch middleware.