feat: Improve CAIP-25 builder error handling
The CAIP-25 builder will now re-throw unexpected errors when validating
whether a given chain ID is supported. Previously when validating a
chain ID, it would suppress the error without checking what it was,
under the assumption that the chain ID was not found in state. This
could be hiding a different error.

The expected error has been moved to a constant that is exported from
the `@metamask/network-controller` package.