Release 312.0.0 (#5399)
Release to update/re-adapt the logic of
`AccountsController:listMultichainAccounts`. We used to rely on the
`account.type` to do some filtering (this was before the introduction of
`account.scopes`).

The logic has been adapted to use the `account.scopes`, which is more
robust and **correct** implementation.

There should not be any impacts for existing accounts with the exception
of Bitcoin accounts that uses different scopes for mainnet and testnet
(Bitcoin support has been disabled temporarily on the extension).