fix: Add MANTLE support for conversion rates (#5402)
## Explanation

Inspecting extension shows that we fetch the conversion rate using this
URL:
https://min-api.cryptocompare.com/data/price?fsym=ETH&tsyms=MNT

This is incorrect as it is not using the correct symbol override. The
url should be:
https://min-api.cryptocompare.com/data/price?fsym=ETH&tsyms=MANTLE

This gives a much better unsupported conversion number that what it
currently is.

## References

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controllers`

- **FIXED**: update `crypto-compare-service` `fetchExchangeRate` to
allow symbol overrides for mantle.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes