Add `getAssetMetadata` action to `MultichainAssetsController` (#5430)
## Explanation

This adds an action to the `MultichainAssetsController` called
`getAssetMetada` which allows to get a specific asset Metadata.

This allows us to get a specific metadata without having to pull the
entire controller state.

## References

* Related to https://github.com/MetaMask/snaps/pull/3166

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/assets-controller`

- **ADDED**: Add `getAssetMetadata` action to
`MultichainAssetsController`.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes