Fix since-latest-release + --include-head (#5016)
The package-level `since-latest-release` script is useful when authoring
release candidate PRs for double-checking that all changes to a package
have been properly documented.

Sometimes it is useful to know all of the changes that have occurred for
a package, even those that took place on the release branch itself. For
this purpose the `--include-head` argument can be passed to the script,
e.g.:

```
yarn workspace @metamask/assets-controllers run since-latest-release --include-head -- diff
```

However, this invocation does not work and produces an error instead.

Additionally, if you pass more than one argument after the `--`, like
so:

```
yarn workspace @metamask/assets-controllers run since-latest-release --include-head -- log -p
```

then the script freezes and does not continue.

This commit addresses both of these issues.