Release/320.0.0 (#5451)
## Explanation

This is releasing a minor version update of `assets-controllers` where
we add a new controller to support search & discovery in mobile.

## Changelog

### `@metamask/assets-controllers`
- **ADDED**: Added a new controller in `assets-controllers`,
`TokenSearchDiscoveryDataController`, which supports the new token
search and discovery feature in MetaMask Mobile. See
https://github.com/MetaMask/metamask-mobile/pull/13328 for how this new
controller is used.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes