Disable type-specific checks for jsdoc/check-tag-names
The `jsdoc/check-tag-names` ESLint rule incorrectly marks the
`@property` JSDoc tag as redundant. Although we usually document
properties in object types directly, the `@property` tag is useful to
document properties that arrive via intersections of other types.
Ideally we should update `eslint-plugin-jsdoc` to allow this JSDoc tag
to be used somehow, but in the meantime we have to disable all of the
type-specific checks within `jsdoc/check-tag-names` so we can avoid a
lint error/warning.