fix: Override lower bound gas when gasLimit is defined (#5396)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

On the extension client, a strange behaviour occurrs when a dApp sends a
request with a `gasLimit` lower than the minimum 21000, but the
transaction still is confirmed with a gas of 21000.

This happens because `updateGasProperties` is called inside
`addTransaction`. `updateGasProperties` calls `updateGas` that calls
`getGas` in which if the gas property in the `txParams` is falsy, it
defaults to FIXED_GAS which is `0x5208` (21000). To fix this issue, we
can derive gas from gasLimit during normalization of transaction
parameters when the first is falsy and the second is defined.

Note that transactions correctly failed when the user tweaked the gas
prices using the appropriate modals. This is because [gas is derived
from gasLimit before new gas settings are submitted to the transaction
controller in
`updateTransactionGasFees`](https://github.com/MetaMask/metamask-extension/blob/838e4cb62765a8dafaf2f00d5e77c8593c91e845/ui/pages/confirmations/hooks/useTransactionFunctions.js#L90).

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

* Fixes: https://github.com/MetaMask/metamask-extension/issues/30183

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/transaction-controller`

- **FIXED**: Derive `gas` from `gasLimit` when the first is undefined
and the second is defined.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes