Split up `multichain` package into 3 packages
Not everything in `@metamask/multichain` is strictly multichain-related.
We could rename this package to `@metamask/caip25`, but not everything
is CAIP-25-related, either.

To better highlight the purpose of the different parts of
`@metamask/multichain`, this commit extracts those parts to 3 packages:

- `@metamask/chain-agnostic-permission`: Defines a permission caveat and
  endowment designed to match [CAIP-25][1] and related specifications.
- `@metamask/eip1193-permission-middleware`: Implements JSON-RPC methods
  for managing permissions as described in [EIP-2255][2] and [MIP-2][3].
- `@metamask/multichain-api-middleware`: Implements JSON-RPC methods in
  the [MetaMask Multichain API][4] that manage and utilize
  [chain-agnostic permissions][5].

Note that the existing code in `@metamask/multichain` has not been
touched. We will deprecate and remove this package in a future commit.

[1]: https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-25.md
[2]: https://eips.ethereum.org/EIPS/eip-2255
[3]: https://github.com/MetaMask/metamask-improvement-proposals/blob/main/MIPs/mip-2.md
[4]: https://github.com/MetaMask/metamask-improvement-proposals/blob/main/MIPs/mip-5.md
[5]: https://github.com/MetaMask/chain-agnostic-permission