chore: adds safeguard for keyring being locked or unlocked (#5473)
## Explanation

We are listening for the `KeyringController` `lock & unlock` events in
order to safeguard the request to the snap.

## References

Fixes:
https://github.com/MetaMask/metamask-extension/issues/30259#issuecomment-2721187172

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/multichain-transactions-controller`

- **Adds**: Safeguard for `KeyringController` `lock & unlock` events

### `@metamask/assets-controllers`

- **Adds**: Safeguard for `KeyringController` `lock & unlock` events in
`MultichainBalancesController`

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>