Fix dependency-related constraints (#5464)
 I've noticed recently that it is impossible to add a controller
dependency just to `dependencies`; one of our existing constraints
demands that it also be added to `peerDependencies`. This is obviously
incorrect.

This commit addresses this (and other related constraints) by the
following:

- Remove `expectControllerDependenciesListedAsPeerDependencies`. This is
the faulty constraint referred to above.
- Correct `expectDependenciesNotInBothProdAndDev` (now
`expectDependenciesNotInBothProdAndDevOrPeer`) to not only ensure that a
dependency is not listed in both `dependencies` and `devDependencies`
but also both `dependencies` and `peerDependencies` (only
`devDependencies` + `peerDependencies` is allowed, as combinations go).
- Add `expectPeerDependenciesAlsoListedAsDevDependencies` constraint to
ensure that peer dependencies are also listed as dev dependencies.
- Fix `expectUpToDateWorkspaceDependenciesAndDevDependencies` to skip
validation of "production" dependencies that are also listed in peer
dependencies (this avoids conflicting constraints).
- Correct `expectConsistentDependenciesAndDevDependencies` to skip
workspace dependencies, as that is already being handled by
`expectUpToDateWorkspaceDependenciesAndDevDependencies` (this avoids
conflicting constraints).

Note that because of these changes, several controllers have also been
corrected to fit.