Split up `multichain` package into 3 packages (#5476)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

Not everything in `@metamask/multichain` is strictly multichain-related.
We could rename this package to `@metamask/caip25`, but not everything
is CAIP-25-related, either.

To better highlight the purpose of the different parts of
`@metamask/multichain`, this commit splits it up into 3 packages:

- `@metamask/chain-agnostic-permission`: Defines an endowment type
permission designed to persist the account and chain components of a
[CAIP-25](https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-25.md)
request. This package also includes adapters and utility functions for
interfacing with this permission.
- `@metamask/eip1193-permission-middleware`: Implements the JSON-RPC
methods for managing permissions as referenced in
[EIP-2255](https://eips.ethereum.org/EIPS/eip-2255) and
[MIP-2](https://github.com/MetaMask/metamask-improvement-proposals/blob/main/MIPs/mip-2.md),
but adapted to support [chain-agnostic permission
caveats](https://npmjs.com/package/@metamask/chain-agnostic-permission).
- `@metamask/multichain-api-middleware`: JSON-RPC methods and middleware
to support the the [MetaMask Multichain
API](https://github.com/MetaMask/metamask-improvement-proposals/blob/main/MIPs/mip-5.md).

Note that the existing code in `@metamask/multichain` has not been
touched. We will deprecate and remove this package later.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

Closes https://github.com/MetaMask/MetaMask-planning/issues/4395.

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

(No updates to changelogs needed)

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>