fix(multichain-transactions-controller): filters out non-mainnet Solana transactions from `:accountTransactionsUpdated` events (#5497)
## Explanation

We were not applying the filtering for Solana mainnet transaction in the
update method `handleOnAccountTransactionsUpdated`, and this was causing
Devnet transactions appearing in the extension.

## References

* Fixes https://github.com/MetaMask/metamask-extension/issues/31072

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/multichain-transactions-controller`

- **Fix**: Applies the mainnet filtering in
`handleOnAccountTransactionsUpdated`


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Charly Chevalier <charly.chevalier@consensys.net>