Expose messenger events & config for RPC failover (#5492)
This commit adds two distinct but related sets of changes to
NetworkController to further support the RPC failover work:

- It allows clients to customize RPC services — which will ultimately be
used to hit RPC endpoints — by passing an `getRpcServiceOptions` option
to the NetworkController constructor. Different RPC endpoints can have
different options, so for instance, one could specify a
`circuitBreakDuration` for all Quicknode RPC endpoints which would be
different for all Infura endpoints.
- It allows clients to tap into key events via the NetworkController
messenger. Namely, the `NetworkController:rpcEndpointUnavailable` event
will be published when too many consecutive failed requests are made to
an RPC endpoint, and `NetworkController:rpcEndpointDegraded` will be
published when an RPC endpoint responds successfully but too slowly.
(There is also a `NetworkController:rpcEndpointRequestRetried` event,
but that is mainly useful for testing.)