Expose error in NetworkController:rpcEndpointUnavailable event (#5501)
In the clients we want the ability to decide whether or not to create a
Segment error when a request to Infura fails too many times in a row.
Specifically, we want to be able to ignore connection errors, as they
may indicate that the user is offline, in which case we don't need to
fail over to Quicknode.

To accommodate this need:

- Expose the error that was encountered while making the request in the
`NetworkController:rpcEndpointUnavailable` event data.
- Export `isConnectionError` from `RpcService`.