fix: gas estimation for type 4 transactions (#5519)
## Explanation

The gas for type 4 transactions that include `data` to the upgraded
account can only be estimated using `eth_estimateGas` if the real
signature properties are included, otherwise the upgraded address cannot
be known since it is derived from the signature itself.

As we don't want to sign an authorisation until the user approves the
transaction, we will instead:

1. Estimate only the upgrade, with no data and a dummy signature, using
`eth_estimateGas`.
2. Estimate the data only on the resulting upgraded EOA using the
simulation API to override the account code.
3. Add the two values together, and subtract the intrinsic gas cost
(`21000`).

## References

Fixes
[#31140](https://github.com/MetaMask/metamask-extension/issues/31140)

## Changelog

See `CHANGELOG.md`.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes