feat: Add messenger `delegate` and `revoke` methods
The methods `delegate` and `revoke` have been added to the Messenger
class. These methods replace the need for the `RestrictedMessenger`.
The `getRestricted` method, and the `RestrictedMessenger` class, have
been removed as obsolete.

See this ADR PR for details: https://github.com/MetaMask/decisions/pull/53

Changelog update still pending, this is a work in progress.