Release/342.0.0 (#5538)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
This release updates some of the method arguments for the
`earn-controller` and updates the `selectedAccountChange` subscription
handler to pass the `account?.address` from the event payload to the
underlying methods.

See changelog for further details.

---------

Co-authored-by: cryptodev-2s <109512101+cryptodev-2s@users.noreply.github.com>