NetworkController: Make failover URLs optional (#5561)
There are two new non-optional properties that were added in a previous
commit:

- `failoverUrls` in `RpcEndpoint` (which is used in NetworkController
state)
- `failoverRpcUrls` in `NetworkClientConfiguration` (which is exposed
via the network client)

Making these properties non-optional makes sense to avoid `undefined`,
however, this is requiring clients that wish to upgrade to implement a
migration to ensure that state is valid along with many other changes to
test files and such which expect state to be a certain shape. We would
rather avoid this right now due to some other changes to
`network-controller` that we need to adapt to, so to make the
maintenance burden lighter, we are changing these to optional. We will
change them back to non-optional in another commit.