refactor(accounts-controller): use type-cast instead of `@ts-expect-error` (#5568)
## Explanation

Commenting out the `@ts-expect-error` was not very reliable sometimes
(it works everytime on the CI, but not everytime when compiling/running
tests locally). We sometimes have to uncomment the directive to run the
tests (which is a bit annoying and force you to have some "unstaged
change" in `git`).

The type-cast does get rid of the compile-error with `WritableDraft`
while still providing the same "object signature" and has 0 impact at
runtime either. So this is safe.

## References

N/A

## Changelog

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes