feat: add backup and sync granular feature control (#5636)
## Explanation

This PR adds a new public method, `setIsBackupAndSyncFeatureEnabled` to
`UserStorageController`
This replaces `enableProfileSyncing` and `disableProfileSyncing` and
will be used as the main method to enable and disable backup and sync
features from now on. As a result, this is a **BREAKING** change.

This PR also adds the `isAccountSyncingEnabled` state property to
`UserStorageController`. This defaults to `true`.

## References

Fixes: https://consensyssoftware.atlassian.net/browse/IDENTITY-81
Extension test drive PR (CI ✅):
https://github.com/MetaMask/metamask-extension/pull/31939

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes