NetworkController: Add flag for controlling RPC failover behavior (#5668)
Recently we added some new behavior to NetworkController which will
detect if an Infura RPC endpoint is down and allow for forwarding
traffic to Quicknode automatically. Because this behavior involves
integrating with backend services to enable Quicknode on failure, we are
unsure whether everything will work the way we want to on the first try,
so we want to be able to disable or enable the behavior using feature
flags. That is, when turned off, even if an RPC endpoint has been
configured with failover URLs, network traffic will not be forwarded to
the failovers if the primary is perceived to be down.

There are a couple of things to note about this commit:

- Usually we would have NetworkController listen to
RemoteFeatureFlagController to know when a feature flag was enabled or
disabled and then enable or disable the RPC failover behavior
accordingly. Because we don't intend to keep the feature flag forever,
however, we let the clients do that and we simply expose methods in
NetworkController that can be used manually by the clients.
- The way that the enabling/disabling works is a bit tricky. Naively, we
could create new versions of network client objects that have different
sets of RpcService objects, and replace the existing versions directly
in the network client registry (without updating their IDs). That's a
little weird, though, and it also means that if another controller
caches a network client at any point in time, the cached version would
not be updated. So instead, we take advantage of the fact that when
other controllers interact with a network client, it's really a proxy —
an AutoManagedNetworkClient — and so we can replace the underlying
object (a NetworkClient) without changing the proxy.