fix: remove metadata for unsupported keyrings (#5725)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
When the user vault is decrypted and there is an attempt to restore an
unsupported/deprecated/faulty keyring there's no mechanism to remove
related metadata, which leads to a situation where no further action can
be made on the controller, because checks for keyrings and metadata
length will fail.

We could remove the related metadata object when the keyring restore
fails, but then we would lose the original ID generated for the keyring.
We can, instead, change the place where the metadata is stored from a
state property to the encrypted vault: by placing the metadata along
with its serialised keyring in the vault we can guarantee a 1:1 link
between them while being able to keep metadata for unsupported keyrings.

Given that we don't need to use the KeyringController state to persist
metadata anymore (as it is persisted along with the vault), we can also
remove `keyringsMetadata` completely, and add a `metadata` attribute to
each keyring in `state.keyrings` instead - which won't be persisted, as
it will be recreated at runtime every time the vault is decrypted and
the keyrings are deserialised.

## References

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->
* Fixes https://github.com/MetaMask/core/issues/5701

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>