chore: Remove obsolete workaround
An error case was added to our network middleware long ago to
workaround load balancer errors that we encountered with Infura at the
time. These errors were fixed long ago, so this workaround is no longer
needed.

I checked with the Infura team, and they confirmed that this case
should no longer be possible for Infura RPC endpoints.

Removing this check allowed me to update how we're parsing the response
body as well. We're now using `response.json()` rather than parsing the
raw body as text. As a consequence of this, we no longer have the raw
text to attach to parsing errors, but this seems OK to remove given that
we don't reference it anywhere, and the full response can be seen in
devtools in a development environment.