chore: Remove obsolete workaround (#5808)
## Explanation

An error case was added to our network middleware long ago to workaround
load balancer errors that we encountered with Infura at the time. These
errors were fixed long ago, so this workaround is no longer needed.

I checked with the Infura team, and they confirmed that this case should
no longer be possible for Infura RPC endpoints.

Removing this check allowed me to update how we're parsing the response
body as well. We're now using `response.json()` rather than parsing the
raw body as text. As a consequence of this, we no longer have the raw
text to attach to parsing errors, but this seems OK to remove given that
we don't reference it anywhere, and the full response can be seen in
devtools in a development environment.

## References

This workaround was originally introduced here:
https://github.com/MetaMask/eth-json-rpc-infura/blame/7871c8ee5acf6c738b6bfa43dfaadc02d7f00509/src/index.js#L13C52-L13C59

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [x] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes