Add `@metamask/error-reporting-service`
This package makes it possible for any module to report an error to an
error reporting app (such as Sentry). The exact mechanism to do so is
customizable, and the service object exposes a messenger action so that
modules can report the error without needing direct access to the
service object itself.