feat(accounts-controller): Add `entropySource` to new `InternalAccount` (#5841)
## Explanation

`AccountsController:accountAdded` events were missing the
`.options.entropySource` property, causing account syncing to misbehave,
registering new accounts to the primary SRP instead of the actual SRP
used to add the account.

We now add the `entropySource` to these new `InternalAccount`s based on
the `keyring` that was used to create them.

## References
Relates to #5618
Relates to #5725
Relates to #5753