Fix block tracker not resolving for failed requests (#5860)
Currently, if a consumer makes a request to an RPC endpoint via a
network provider object, and that request continually fails in some way
and then throws a circuit breaker error, it will never resolve.
Naturally, this causes problems if the consuming code (e.g.
TransactionController, or UI code) is expecting a response from the
network in order to proceed.

This problem occurs because internally the NetworkController wraps
requests in special middleware that use the block tracker to get the
latest block number. This middleware gets executed before the desired
request reaches the network, and in this case because the block tracker
does not return a resolved promise, the middleware does not resolve
along with the entire request.

This bug has been resolved in `@metamask/eth-block-tracker` 12.0.1 and
`@metamask/eth-json-rpc-middleware` 17.0.1. So, to fix this problem,
this PR upgrades these packages.