fix(accounts-controller): fix Snap account's metadata update when Snaps during wallet reset flow (#5884)
## Explanation

We assume the `snap` will always be defined, but they could be
deleted/reloaded during some flows. To prevent using an `undefined`
`snap` here, we check if it's defined first, if not, we consider it to
be disabled.

## References

- https://github.com/MetaMask/metamask-mobile/issues/15628

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes