chore: explicitly pass networkClientId for nftController (#5622)
## Explanation

This PR is part of a larger effort to remove the global network
selector.

On this PR:
* Removal of this.#chainId on the nftController
* Removal of chainId from nftController constructor
* Making the optional networkClientId in all functions a mandatory field

UI integration:
- extension: https://github.com/MetaMask/metamask-extension/pull/31917

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

* Fixes #5581 

<!--
Are there any issues that this pull request is tied to?
Are there other links that reviewers should consult to understand these
changes better?
Are there client or consumer pull requests to adopt any breaking
changes?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
THIS SECTION IS NO LONGER NEEDED.

The process for updating changelogs has changed. Please consult the
"Updating changelogs" section of the Contributing doc for more.
-->

### `@metamask/assets-controllers`

- **UPDATE:** Updated `NftController` and `NftDetectionController` to
eliminate the dependency on the current chain. All functions that
previously accepted networkClientId as an optional parameter now require
it as a mandatory parameter

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've communicated my changes to consumers by [updating changelogs
for packages I've
changed](https://github.com/MetaMask/core/tree/main/docs/contributing.md#updating-changelogs),
highlighting breaking changes as necessary
- [ ] I've prepared draft pull requests for clients and consumer
packages to resolve any breaking changes

---------

Co-authored-by: salimtb <salim.toubal@outlook.com>