Add DeepWiki badge
DeepWiki is a site powered by Devin that analyzes a repo and generates a
comprehensive breakdown of the architecture. (The one for this repo is
located [here](https://github.com/MetaMask/core).) It doesn't always
organize information in a way that humans would, but it still does a
pretty good job of summarizing and finding connections between modules.
Deepwiki will automatically update its report as long as we have this
badge in the README.